+++ index.html
@@ -230,7 +230,7 @@
     </div>
   </template>
 
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
   <script src="https://lodash.com/docs/4.17.15#random"></script>
 </body>
 </html>
